Updated bin/sandbox
Adds $PAY_TOMORROW_REDIRECT_BASE_URL to config.hosts if present.
This variable is required for paytomorrow redirects.

In dev, this var holds the ngrok url and the step to add it
to config.hosts is automated with the sed command

If the variable is not present, info to set it is logged