[Bug fix] Don't call authorize after confirm->complete step
Solidus assumes payments in checkout state as unprocessed_payments and
calls purchase when auto-capture is enabled and
authorize when auto_capture is disabled.

To enable this -
The payment should be in checkout state when auto_capture is enabled.
So that purchase is already called &
The payment should be in pending state when auto_Capture is disabled.
So that authorize is NOT called